requirements.txt
mss
google-genai
edge-tts
pygame
python-dotenv
pillow
cffi
sounddevice
numpy
